adjusted names for noshirt and ohlc classes to new names(StrategyManager and CandleData)
Also cleaned app.py file removing duplicate code implemented by tradingBot.py
